requirements.txt
ultralytics
opencv-python
picamera2
PyYAML
